Move permutation matrix Mul impl to its own file
